pyproject.toml
[tool.poetry]
name = "latexstats"
version = "0.1.0"
description = "Upload stats about latex documents to the stats database"
authors = ["George Kaye <georgejkaye@gmail.com>"]
license = "GPL-v3.0-or-later"
readme = "README.md"
packages = [{include = "latexstats", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
